delete KV

+++ deploy/purebicep/cleanup.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+set +e
 AZHOP_CONFIG=config.yml
 ANSIBLE_VARIABLES=playbooks/group_vars/all.yml
 
@@ -9,10 +10,12 @@ fi
 
 rg=$(yq '.resource_group' $AZHOP_CONFIG)
 
-# Purge keyvault
+# Delete and Purge keyvault
 key_vault=$(yq eval '.key_vault' $ANSIBLE_VARIABLES)
 if [ "$key_vault" != "" ]; then
-  echo "Purging keyvault $key_vault"
+    echo "Delete keyvault $key_vault"
+    az keyvault delete --name $key_vault
+    echo "Purging keyvault $key_vault"
     az keyvault purge --name $key_vault
 fi
 
@@ -27,6 +30,11 @@ fi
 peered_vnet=$(yq ".network.peering[0].vnet_name" $AZHOP_CONFIG)
 if [ "$peered_vnet" != "null" ]; then
     peered_vnet_rg=$(yq ".network.peering[0].vnet_resource_group" $AZHOP_CONFIG)
-    id=$(az network vnet peering list -g $peered_vnet_rg --vnet-name $peered_vnet --query "[?remoteVirtualNetwork.resourceGroup=='$rg'].id" -o tsv)
-    az network vnet peering delete --ids $id
+    if [ "$peered_vnet_rg" != "null" ]; then
+      id=$(az network vnet peering list -g $peered_vnet_rg --vnet-name $peered_vnet --query "[?remoteVirtualNetwork.resourceGroup=='$rg'].id" -o tsv)
+      if [ "$id" != "" ]; then
+        echo "Deleting peering $id"
+        az network vnet peering delete --ids $id
+      fi
+    fi
 fi
